+++ python-package/geobr/read_amazon.py
@@ -1,5 +1,7 @@
 from geobr.utils import read_geobr_v2
+from geobr._docstrings import docparams
 
+@docparams
 def read_amazon(
     year: int,
     simplified: bool = True,
@@ -12,10 +14,13 @@ def read_amazon(
 
     Parameters
     ----------
-    year : int
-        Year of the data.
-    simplified, verbose, output, show_progress, cache
-        Standard geobr options.
+    {year}
+    {simplified}
+    {verbose}
+    {output}
+    {show_progress}
+    {cache}
+
     """
 
     return read_geobr_v2(