+++ python-package/geobr/read_country.py
@@ -1,6 +1,8 @@
 from geobr.utils import read_geobr_v2
+from geobr._docstrings import docparams
 
 
+@docparams
 def read_country(
     year: int,
     simplified: bool = True,
@@ -13,12 +15,15 @@ def read_country(
 
     Parameters
     ----------
-    year : int
-        Year of the data.
-    simplified, verbose, output, show_progress, cache
-        Standard geobr options.
+    {year}
+    {simplified}
+    {verbose}
+    {output}
+    {show_progress}
+    {cache}
+
     """
-    
+
     return read_geobr_v2(
         "country",
         year,